+++ CHANGELOG.md
@@ -7,6 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added Get-TkMsalToken cmdlet to retrieve an MSAL token using API calls.
+- Added Managed Identity support for  Get-TkMsalToken cmdlet (Needs to be tested).
+- SecureString support for Get-TkMsalToken cmdlet.
+- Formatting alignment for cmdlets.
+
+### Fixed
+
+- Fixed authentication context for MgGraph.
+
+### Changed
+
+- Updated private function names to be more descriptive.
+- Removed MSAL.PS dependency from Send-TkEmailAppMessage function.
+- Removed Send-TkEmailAppMessage module install if manual parameters are provided.
+
+## [0.2.0] - 2025-03-14
+
+### Added
+
 - Updated docs for the module.
 - Release Candidate
 - Update wiki pages.